Allow simplejpeg to be linked against older versions of libjpeg-turbo
Older versions of libjpeg-turbo (such as the one packaged with Debian
Bookworm) don't have the TJSAMP_UNKNOWN symbol, so replacing its
occurrences in simplejpeg by -1 (which happens to be its true value)
makes it possible to link against those older libraries.

There should be no functional difference, and this change should be
regarded as temporary until such time as newer versions of
libjpeg-turbo are more widespread.